feat(calendar): implement event handler for create calendar on user created event handler and refactor some files names
