+++ src/infra/crqs/calendar/calendar.module.ts
@@ -5,13 +5,16 @@ import { CreateCalendarCommandHandler } from '@infra/crqs/calendar/commands/crea
 import { FetchUserCalendarWithRowsQueryHandler } from '@infra/crqs/calendar/queries/fetch-user-calendar-with-rows.query';
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
+import { CreateCalendarOnNewUserHandler } from '@infra/crqs/calendar/handlers/create-calendar-on-new-user.handler';
 
 const CommandHandlers = [CreateCalendarCommandHandler, AddRowInCalendarCommandHandler];
 
 const QueryHandlers = [FetchUserCalendarWithRowsQueryHandler];
 
+const EventHandlers = [CreateCalendarOnNewUserHandler];
+
 @Module({
   imports: [CqrsModule.forRoot()],
-  providers: [CreateCalendar, AddWorkInCalendar, ...CommandHandlers, ...QueryHandlers],
+  providers: [CreateCalendar, AddWorkInCalendar, ...CommandHandlers, ...QueryHandlers, ...EventHandlers],
 })
 export class CalendarModule {}
