+++ src/infra/crqs/notification/handlers/on-user-created.ts
@@ -1,6 +1,6 @@
-import { UserCreated } from '@domain/auth/application/events/user-created';
 import { EventsHandler, IEventHandler } from '@nestjs/cqrs';
 import { CreateSubscriber } from '@domain/notifications/application/use-cases/create-subscriber';
+import { UserCreated } from '@domain/auth/enterprise/events/user-created';
 
 @EventsHandler(UserCreated)
 export class OnUserCreatedEventHandler implements IEventHandler<UserCreated> {
