+++ src/infra/database/prisma/prisma-calendar.repository.ts
@@ -1,6 +1,6 @@
 import { CalendarRepository } from '@domain/calendar/application/contracts/calendar-repository';
-import { Calendar } from '@domain/calendar/enterprise/entities/calendar';
-import { CalendarRow } from '@domain/calendar/enterprise/entities/calendar-row';
+import { Calendar } from '@domain/calendar/enterprise/entities/Calendar';
+import { CalendarRow } from '@domain/calendar/enterprise/entities/Calendar-row';
 import {
   parsePrismaCalendarRowToDomainCalendarRow,
   parsePrismaCalendarToDomainCalendar,
