+++ test/mocks/in-memory-calendar-repository.ts
@@ -1,6 +1,6 @@
 import { CalendarRepository } from '@domain/calendar/application/contracts/calendar-repository';
-import { Calendar } from '@domain/calendar/enterprise/entities/calendar';
-import { CalendarRow } from '@domain/calendar/enterprise/entities/calendar-row';
+import { Calendar } from '@domain/calendar/enterprise/entities/Calendar';
+import { CalendarRow } from '@domain/calendar/enterprise/entities/Calendar-row';
 import { filter, find } from 'lodash';
 
 export class InMemoryCalendarRepository implements CalendarRepository {
